refactor: Move API item utils into directory
