+++ tools/api-markdown-documenter/src/api-item-transforms/TransformApiItem.ts
@@ -28,13 +28,13 @@ import type { ApiDocument } from "../ApiDocument.js";
 import type { Section } from "../mdast/index.js";
 import { getApiItemKind } from "../utilities/index.js";
 
-import {
-	doesItemRequireOwnDocument,
-	shouldItemBeIncluded,
-} from "./ApiItemTransformUtilities.js";
 import type { ApiItemTransformationConfiguration } from "./configuration/index.js";
 import { createBreadcrumbParagraph } from "./helpers/index.js";
-import { createDocument } from "./utilities/index.js";
+import {
+	createDocument,
+	doesItemRequireOwnDocument,
+	shouldItemBeIncluded,
+} from "./utilities/index.js";
 
 /**
  * Creates a {@link ApiDocument} for the specified `apiItem`.
