+++ tools/api-markdown-documenter/src/api-item-transforms/TransformApiModel.ts
@@ -14,10 +14,6 @@ import {
 import type { ApiDocument } from "../ApiDocument.js";
 import type { Section } from "../mdast/index.js";
 
-import {
-	doesItemRequireOwnDocument,
-	shouldItemBeIncluded,
-} from "./ApiItemTransformUtilities.js";
 import { apiItemToDocument, apiItemToSections } from "./TransformApiItem.js";
 import {
 	type ApiItemTransformationConfiguration,
@@ -25,7 +21,12 @@ import {
 	getApiItemTransformationConfigurationWithDefaults,
 } from "./configuration/index.js";
 import { createBreadcrumbParagraph, createEntryPointList } from "./helpers/index.js";
-import { checkForDuplicateDocumentPaths, createDocument } from "./utilities/index.js";
+import {
+	checkForDuplicateDocumentPaths,
+	createDocument,
+	doesItemRequireOwnDocument,
+	shouldItemBeIncluded,
+} from "./utilities/index.js";
 
 /**
  * Renders the provided model and its contents to a series of {@link ApiDocument}s.
