+++ tools/api-markdown-documenter/src/api-item-transforms/default-implementations/TransformApiModuleLike.ts
@@ -18,9 +18,9 @@ import {
 import type { Section } from "../../mdast/index.js";
 import type { ApiModuleLike } from "../../utilities/index.js";
 import { getApiItemKind, getScopedMemberNameForDiagnostics } from "../../utilities/index.js";
-import { filterItems } from "../ApiItemTransformUtilities.js";
 import type { ApiItemTransformationConfiguration } from "../configuration/index.js";
 import { createChildDetailsSection, createMemberTables } from "../helpers/index.js";
+import { filterItems } from "../utilities/index.js";
 
 /**
  * Default documentation transform for module-like API items (packages, namespaces).
