+++ tools/api-markdown-documenter/src/api-item-transforms/default-implementations/TransformApiTypeLike.ts
@@ -21,9 +21,9 @@ import {
 	isStatic,
 	type ApiTypeLike,
 } from "../../utilities/index.js";
-import { getFilteredMembers } from "../ApiItemTransformUtilities.js";
 import type { ApiItemTransformationConfiguration } from "../configuration/index.js";
 import { createChildDetailsSection, createMemberTables } from "../helpers/index.js";
+import { getFilteredMembers } from "../utilities/index.js";
 
 /**
  * Default documentation transform for {@link ApiTypeLike | type-like} API items.
