+++ tools/api-markdown-documenter/src/api-item-transforms/index.ts
@@ -15,7 +15,7 @@ export {
 	getHeadingForApiItem,
 	getLinkForApiItem,
 	shouldItemBeIncluded,
-} from "./ApiItemTransformUtilities.js";
+} from "./utilities/index.js";
 export {
 	type ApiItemTransformationConfiguration,
 	type ApiItemTransformationConfigurationBase,
