+++ tools/api-markdown-documenter/src/api-item-transforms/utilities/ApiItemTransformUtilities.ts
@@ -8,15 +8,14 @@ import { strict as assert } from "node:assert";
 import { type ApiItem, ApiItemKind } from "@microsoft/api-extractor-model";
 import type { Link } from "mdast";
 
-import type { SectionHeading } from "../mdast/index.js";
+import type { SectionHeading } from "../../mdast/index.js";
 import {
 	getApiItemKind,
 	getFilteredParent,
 	getFileSafeNameForApiItem,
 	type ValidApiItemKind,
 	getEffectiveReleaseLevel,
-} from "../utilities/index.js";
-
+} from "../../utilities/index.js";
 import {
 	FolderDocumentPlacement,
 	HierarchyKind,
@@ -25,7 +24,7 @@ import {
 	type FolderHierarchyConfiguration,
 	type DocumentationHierarchyConfiguration,
 	type HierarchyConfiguration,
-} from "./configuration/index.js";
+} from "../configuration/index.js";
 
 /**
  * This module contains `ApiItem`-related utilities for use in transformation logic.
