+++ tools/api-markdown-documenter/src/api-item-transforms/utilities/index.ts
@@ -3,5 +3,16 @@
  * Licensed under the MIT License.
  */
 
+export {
+	createQualifiedDocumentNameForApiItem,
+	doesItemRequireOwnDocument,
+	doesItemKindRequireOwnDocument,
+	filterItems,
+	getFilteredMembers,
+	getHeadingForApiItem,
+	getLinkForApiItem,
+	isItemOrAncestorExcluded,
+	shouldItemBeIncluded,
+} from "./ApiItemTransformUtilities.js";
 export { createDocument, checkForDuplicateDocumentPaths } from "./DocumentUtilities.js";
 export { resolveSymbolicLink } from "./ReferenceUtilities.js";
